Verify login functionality

+++ bmsrestfulapi/src/main/java/com/bmsrestfulapi/services/UserServiceImpl.java
@@ -13,6 +13,7 @@ import com.bmsrestfulapi.entities.User;
 import com.bmsrestfulapi.exceptions.CustomExceptionsMessages;
 import com.bmsrestfulapi.exceptions.EmptyUserListException;
 import com.bmsrestfulapi.exceptions.InvalidCredentialsException;
+import com.bmsrestfulapi.exceptions.NotLoggedInException;
 import com.bmsrestfulapi.exceptions.UserNotCreatedException;
 import com.bmsrestfulapi.exceptions.UserNotFoundException;
 import com.bmsrestfulapi.repositories.AccountInfoRepository;
@@ -55,6 +56,7 @@ public class UserServiceImpl implements UserService {
 			if (user.getGender().equalsIgnoreCase("male") || user.getGender().equalsIgnoreCase("female")) {
 				user.setName(user.getName().toLowerCase());
 				user.setGender(user.getGender().toLowerCase());
+				user.setAddress(user.getAddress().toLowerCase());
 				User u = userRepository.save(user);
 				u.getLogin().setAccountNo(u.getAccountList().get(0).getAccountNo());
 				return "User created Successfully\nDetails:\n" + u;
@@ -67,14 +69,19 @@ public class UserServiceImpl implements UserService {
 
 	@Override
 	public String verifyUser(Integer userId, Integer adminId)
-			throws InvalidCredentialsException, UserNotFoundException {
+			throws InvalidCredentialsException, UserNotFoundException, NotLoggedInException {
 		if (userRepository.existsById(adminId)) {
 			User admin = userRepository.getById(adminId);
 			if (admin.getRole().getRoleName().equalsIgnoreCase(UserService.ADMIN)) {
 				if (userRepository.existsById(userId)) {
-					Login login = loginRepository.getLoginById(userId);
-					login.setVerified(true);
-					return " User verified Successfully.";
+					if (admin.getLogin().isLogin()) {
+						Login login = loginRepository.getLoginById(userId);
+						login.setVerified(true);
+						return " User verified Successfully.";
+					} else {
+						throw new NotLoggedInException(CustomExceptionsMessages.YOU_ARE_NOT_LOGGEDIN);
+					}
+
 				} else {
 					throw new UserNotFoundException(CustomExceptionsMessages.NO_USER_EXISTS_WITH_THIS_ID);
 				}
@@ -87,16 +94,21 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public String checkBalance(Integer userId, Integer pin) throws InvalidCredentialsException {
+	public String checkBalance(Integer userId, Integer pin) throws InvalidCredentialsException, NotLoggedInException {
 		User user = userRepository.getPin(pin, userId);
 		if (user != null && user.getPin().equals(pin) && user.getUserId().equals(userId)) {
-			return "Your current balance is: " + accountInfoRepository.getBalance(userId);
+			if (user.getLogin().isLogin()) {
+				return "Your current balance is: " + accountInfoRepository.getBalance(userId);
+			} else {
+				throw new NotLoggedInException(CustomExceptionsMessages.YOU_ARE_NOT_LOGGEDIN);
+			}
 		} else
 			throw new InvalidCredentialsException(CustomExceptionsMessages.PLEASE_CHECK_YOUR_PIN_AND_USER_ID);
 	}
 
 	@Override
-	public String withdrawMoney(Integer accountNo, Integer amount, Integer pin) throws InvalidCredentialsException {
+	public String withdrawMoney(Integer accountNo, Integer amount, Integer pin)
+			throws InvalidCredentialsException, NotLoggedInException {
 		AccountInfo accountInfo = accountInfoRepository.verifyPin(pin, accountNo);
 		if (accountInfo == null) {
 			throw new InvalidCredentialsException(CustomExceptionsMessages.INVALID_CREDENTIAL);
@@ -107,10 +119,15 @@ public class UserServiceImpl implements UserService {
 				&& accountInfo.getUser().getUserId().equals(user.getUserId())) {
 			Integer availableBalance = accountInfo.getCurrentBalance();
 			if (amount <= availableBalance) {
-				availableBalance -= amount;
-				accountInfo.setCurrentBalance(availableBalance);
-				accountInfoRepository.save(accountInfo);
-				return "Money Withdrawn Successfully! \nYour available balance is: " + availableBalance;
+				if (user.getLogin().isLogin()) {
+					availableBalance -= amount;
+					accountInfo.setCurrentBalance(availableBalance);
+					accountInfoRepository.save(accountInfo);
+					return "Money Withdrawn Successfully! \nYour available balance is: " + availableBalance;
+				} else {
+					throw new NotLoggedInException(CustomExceptionsMessages.YOU_ARE_NOT_LOGGEDIN);
+				}
+
 			} else {
 				return "Insufficient Balance";
 			}
@@ -121,7 +138,7 @@ public class UserServiceImpl implements UserService {
 
 	@Override
 	public String moneyTransfer(Integer accountNo, Integer receiversAccountNo, Integer amount, Integer pin)
-			throws InvalidCredentialsException {
+			throws InvalidCredentialsException, NotLoggedInException {
 		AccountInfo accountInfo = accountInfoRepository.verifyPin(pin, accountNo);
 		if (accountInfo == null) {
 			throw new InvalidCredentialsException(CustomExceptionsMessages.INVALID_CREDENTIAL);
@@ -136,14 +153,19 @@ public class UserServiceImpl implements UserService {
 			Integer availableBalance = accountInfo.getCurrentBalance();
 			Integer receiversAvailableBalance = receiverAccountInfo.getCurrentBalance();
 			if (amount <= availableBalance) {
-				availableBalance -= amount;
-				accountInfo.setCurrentBalance(availableBalance);
-				accountInfoRepository.save(accountInfo);
+				if (user.getLogin().isLogin()) {
+					availableBalance -= amount;
+					accountInfo.setCurrentBalance(availableBalance);
+					accountInfoRepository.save(accountInfo);
 
-				receiversAvailableBalance += amount;
-				receiverAccountInfo.setCurrentBalance(receiversAvailableBalance);
-				accountInfoRepository.save(receiverAccountInfo);
-				return "Money Transfered Successfully! \nYour current balance is: " + availableBalance;
+					receiversAvailableBalance += amount;
+					receiverAccountInfo.setCurrentBalance(receiversAvailableBalance);
+					accountInfoRepository.save(receiverAccountInfo);
+					return "Money Transfered Successfully! \nYour current balance is: " + availableBalance;
+				} else {
+					throw new NotLoggedInException(CustomExceptionsMessages.YOU_ARE_NOT_LOGGEDIN);
+				}
+
 			} else {
 				return "Insufficient Balance";
 			}
@@ -154,12 +176,16 @@ public class UserServiceImpl implements UserService {
 
 	@Override
 	public String deleteUserById(Integer userId, Integer adminId)
-			throws UserNotFoundException, InvalidCredentialsException {
+			throws UserNotFoundException, InvalidCredentialsException, NotLoggedInException {
 		User admin = userRepository.getById(adminId);
 		if (admin.getRole().getRoleName().equalsIgnoreCase(UserService.ADMIN)) {
 			if (userRepository.existsById(userId)) {
-				userRepository.deleteById(userId);
-				return "User deleted successfully!";
+				if (admin.getLogin().isLogin()) {
+					userRepository.deleteById(userId);
+					return "User deleted successfully!";
+				} else {
+					throw new NotLoggedInException(CustomExceptionsMessages.YOU_ARE_NOT_LOGGEDIN);
+				}
 			} else {
 				throw new UserNotFoundException(CustomExceptionsMessages.NO_USER_EXISTS_WITH_THIS_ID);
 			}
@@ -169,13 +195,19 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public String updateUser(User user, Integer adminId) throws UserNotFoundException, InvalidCredentialsException {
+	public String updateUser(User user, Integer adminId)
+			throws UserNotFoundException, InvalidCredentialsException, NotLoggedInException {
 		if (userRepository.existsById(adminId)) {
 			User admin = userRepository.getById(adminId);
 			if (admin.getRole().getRoleName().equalsIgnoreCase(UserService.ADMIN)) {
 				if (userRepository.existsById(user.getUserId())) {
-					userRepository.save(user);
-					return "User Updated successfully!";
+					if (admin.getLogin().isLogin()) {
+
+						userRepository.save(user);
+						return "User Updated successfully!";
+					} else {
+						throw new NotLoggedInException(CustomExceptionsMessages.YOU_ARE_NOT_LOGGEDIN);
+					}
 				} else {
 					throw new UserNotFoundException(CustomExceptionsMessages.NO_USER_EXISTS_WITH_THIS_ID);
 				}
@@ -189,13 +221,18 @@ public class UserServiceImpl implements UserService {
 
 	@Override
 	public String getAllUsers(Integer adminId)
-			throws EmptyUserListException, InvalidCredentialsException, UserNotFoundException {
+			throws EmptyUserListException, InvalidCredentialsException, UserNotFoundException, NotLoggedInException {
 		if (userRepository.existsById(adminId)) {
 			User admin = userRepository.getById(adminId);
 			if (admin.getRole().getRoleName().equalsIgnoreCase(UserService.ADMIN)) {
 				List<User> userList = userRepository.findAll();
 				if (!userList.isEmpty()) {
-					return "List of Users: \n" + userList;
+					if (admin.getLogin().isLogin()) {
+						return "List of Users: \n" + userList;
+					} else {
+						throw new NotLoggedInException(CustomExceptionsMessages.YOU_ARE_NOT_LOGGEDIN);
+					}
+
 				} else {
 					throw new EmptyUserListException(CustomExceptionsMessages.NO_USER_EXIST_IN_DATABASE);
 				}
